=add action

+++ app/modules/issue/service.py
@@ -3,14 +3,19 @@ from uuid import UUID
 
 from fastapi import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
+from pydantic import TypeAdapter
 
 from app.infrastructure.db.models import User, Issue
 from app.modules.auth.dependencies import DBSession
+from app.modules.comments.service import CommentService
 from app.modules.issue.repository import IssueRepository
-from app.modules.issue.schema import IssueCreate, IssueUpdate
+from app.modules.issue.schema import IssueCreate, IssueUpdate, IssueResponse, IssueResponseDetail, IssueResponseEdit
 from app.modules.project_members.repository import ProjectMemberRepository
 from app.modules.projects.repository import ProjectRepository
 from app.modules.users.repository import UserRepository
+
+
+ISSUE_LIST_ADAPTER = TypeAdapter(list[IssueResponse])
 
 
 class IssueService:
@@ -28,8 +33,8 @@ class IssueService:
         self.user_repository = user_repository
         self.db = db
 
-    async def create(self,project_id: UUID,data: IssueCreate, current_user: User) -> Issue:
-        project = await self.project_repository.get_by_public_id(self.db,project_id)
+    async def create(self,project_id: UUID,data: IssueCreate, user: User) -> IssueResponse:
+        project = await self.project_repository.get_by_public_id_no_full(self.db, project_id)
 
         if not project:
             raise ValueError("Project not found")
@@ -37,7 +42,7 @@ class IssueService:
         assignee = None
 
         if data.assignee_public_id:
-            assignee = await self.user_repository.get_by_public_id(self.db,data.assignee_public_id)
+            assignee = await self.user_repository.get_by_public_id(self.db, data.assignee_public_id)
 
             if not assignee:
                 raise ValueError("User not found")
@@ -49,7 +54,7 @@ class IssueService:
 
         issue = Issue(
             project_id=project.id,
-            reporter_id=current_user.id,
+            reporter_id=user.id,
             assignee_id=assignee.id if assignee else None,
             title=data.title,
             description=data.description,
@@ -61,26 +66,39 @@ class IssueService:
 
         await self.db.commit()
 
-        return issue
+        return IssueResponse.model_validate(issue)
 
-    async def get_all( self,project_id: UUID) -> list[Issue]:
-        project = await self.project_repository.get_by_public_id(self.db,project_id,)
+    async def get(self, public_id: UUID) -> IssueResponseDetail | None:
+        issue = await self.repository.get_by_public_id_full(self.db, public_id)
+        if issue is None:
+            return None
+
+        base = IssueResponse.model_validate(issue)
+        comments_tree = CommentService.build_comment_tree(issue.comments)
+
+        return IssueResponseDetail(**base.model_dump(), comments=comments_tree)
+
+    async def get_edit(self, public_id: UUID) -> IssueResponseEdit | None:
+        issue = await self.repository.get_by_public_id_edit(self.db, public_id)
+        if issue is None:
+            return None
+        return IssueResponseEdit.model_validate(issue)
+
+    async def list(self,project_id: UUID, user: User, query: str | None = None) -> list[IssueResponse]:
+        project = await self.project_repository.get_by_public_id_no_full(self.db, project_id)
 
         if not project:
             raise ValueError("Project not found")
 
-        return await self.repository.get_all_by_project(self.db,project.id)
+        list_issues = await self.repository.get_all_by_project(self.db, project.id, user.id, query)
 
-    async def get_by_public_id(self,public_id: UUID,) -> Issue:
-        issue = await self.repository.get_by_public_id(self.db,public_id)
+        return ISSUE_LIST_ADAPTER.validate_python(list_issues)
+
+    async def update(self,public_id: UUID,data: IssueUpdate) -> IssueResponse:
+        issue = await self.repository.get_by_public_id(self.db, public_id)
 
         if not issue:
             raise ValueError("Issue not found")
-
-        return issue
-
-    async def update(self,public_id: UUID,data: IssueUpdate) -> Issue:
-        issue = await self.get_by_public_id(public_id)
 
         if data.assignee_public_id is not None:
             assignee = await self.user_repository.get_by_public_id(self.db,data.assignee_public_id)
@@ -95,29 +113,24 @@ class IssueService:
 
             issue.assignee_id = assignee.id
 
-        if data.title is not None:
-            issue.title = data.title
+        update_data = data.model_dump(exclude_unset=True,exclude={"assignee_public_id"})
 
-        if data.description is not None:
-            issue.description = data.description
+        for field, value in update_data.items():
+            setattr(issue, field, value)
 
-        if data.status is not None:
-            issue.status = data.status
-
-        if data.priority is not None:
-            issue.priority = data.priority
-
-        if data.due_date is not None:
-            issue.due_date = data.due_date
-
-        issue = await self.repository.update( self.db,issue)
+        await self.repository.update(self.db, issue)
 
         await self.db.commit()
 
-        return issue
+        issue = await self.repository.get_by_public_id_full(self.db, public_id)
+
+        return IssueResponse.model_validate(issue)
 
     async def delete(self,public_id: UUID) -> None:
-        issue = await self.get_by_public_id(public_id)
+        issue = await self.repository.get_by_public_id(self.db, public_id)
+
+        if not issue:
+            raise ValueError("Issue not found")
 
         await self.repository.delete(self.db, issue)
